refactor form logic to its own hook
